[hphp] fixed server stats related crash in evhttp_ async functions
Summary: Found the problem. (1) ServerStats is thread-local, so logging stats on the new evhttp thread is pointless, as it won’t tally back to main HTTP thread’s. I removed them. (2) because of (1), it tries to create a new ServerStats object, which is added to a vector<> that’s not mutex protected when server status is queried. That can corrupt the data structure. Fixed it by adding a mutex.

Reviewed By: arudolph

Test Plan: make www, makes sure server status works before. waiting for Arthur to tell me if problem is fixed.



git-svn-id: http://svn.facebook.net/hiphop/fbomb/branches/hphp-dev/fbcode/hphp@21422 2248de34-8caa-4a3c-bc55-5e52d9d7b73a