Fix type inference for include
Summary:
Type inference set the actual type for an include statement to bool.
Even though its wrong, we mostly got away with it. But if the include was used
in a larger expression, and the include parameters needed their order enforcing,
it could result in the return value of the include being stored in a bool
variable (or even a byte or char).

% cat bar.php
<?php
return 257;
% cat foo.php
<?php
function f() { return "bar"; }
function g() { return ".php"; }
var_dump($x += require f().g());
% hphp foo.php --ffile=bar.php
int(1)

This appears to be the reason for 400fdd3db96ad8dccae402f1179f9eda18f56321
breaking the Alite site. There was an include that returned an array:

$map = require_once
$_SERVER['PHP_ROOT'].'/lib/alite/site/'.$name.'/urimap.php';

If array accesses are marked as having an effect, the rhs is stored in a temp,
which has type bool. But the require actually returns an array. This then throws
when $map is used where an array is required.

Test Plan:
fast_tests, slow_tests, and the above test. Dont know how to add it to our test
suite.
I have a follow up commit which reverts the revert of
400fdd3db96ad8dccae402f1179f9eda18f56321, and I will also run unit tests and
production tests on that one.

DiffCamp Revision: 117294
Reviewed By: andrewparoski
Commenters: hzhao
CC: hphp-diffs@lists, hzhao, andrewparoski
Revert Plan:
OK