Fix Soap errors introduced by StaticString changes
Summary:
Soap has a map of Variant types to SOAP types, and KindOfStaticString wasnt in
there.
Also, it tries to type arrays based on the types of their elements. For this to
work we have to map StaticString and LiteralString onto String, so that an array
with a mix of elements still gets type String (rather than "mixed").

Test Plan:
fast_tests, scripts/unittest/tests/fbcdn/FbcdnPurgeTestCase.php

DiffCamp Revision: 118696
Reviewed By: myang
CC: hphp-diffs@lists, myang, mwilliams
Revert Plan:
OK