[Fix] ArrayInit needs to clean up in case of exceptions
Summary:
With http://tools.intern.facebook.com/D118453, ArrayInit may be interrupted by
an exception, and therefore the destructor of ArrayInit should not assert the
ArrayData pointer has been consumed, but rather clean it up in case it is not.

Add a new testcase in TestArrayInit.

Test Plan:
TestCodeRun TestArrayInit
Verified that there's no assertion failure in case of exception in the middle.

DiffCamp Revision: 119139
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK