fixed crash in late static binding
Summary:
For some reason, without this change, we generate a piece of code that has
problems in reference counting. I don't want to check in this one, but i want to
talk about this to fully understand what's going on. It has something to do with
the difference between these two pieces of code,

  const String &tmp = <exp>;

vs.

  const String &tmp = (newexp, wrapper->call(<exp>));

The 2nd form was not able to hold tmp with enough ref count.

Test Plan:
compile www with ENABLE_LATE_STATIC_BINDING and run without crash

DiffCamp Revision: 119322
Reviewed By: mwilliams
CC: hphp-diffs@lists, hzhao, mwilliams
Revert Plan:
OK