Make interfaces into first class objects
Summary:
Add Option::UseVirtualDispatch to control interfaces/virtual dispatch. When its
on:
 o Add typedef p_T for SmartInterface<T> type. This is an Object with a pointer
to an Interface, with operator->() overloaded to return the interface.
 o Generate base classes for each interface
 o Declare abstract virtual functions for each abstract method.

For now, UseVirtualDispatch is turned off, because there are a lot of issues to
solve before we can enable it.

Test Plan:
ran fast_tests, slow_tests, built www-dev and ran flib tests.

DiffCamp Revision: 123483
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK