Annotate emitted source with compiler line number
Summary:
Command-line flag --fl-annotate will turn on verbse logging of compiler
file-line info that emits a bit of code.  Note that using
cg_printf in an expression might break the macro (doesn't happen now).

Test Plan:
internal debugging only, testing is awkward.
All that is needed is a manual verification this does nothing when the
option is off, and if it breaks then when on, it doesn't matter.

DiffCamp Revision: 122937
Reviewed By: andrewparoski
CC: hphp-diffs@lists, andrewparoski, ervan
Revert Plan:
OK