[Fix] update source_info with repartitioned files
Summary:

When GenerateSourceInfo is turned on for compilation, a map is generated to map cpp locations to php locations, but it was not accurate after repartitioning the files.

This diff fixes it.

Task ID: #

Blame Rev:

Reviewers: hzhao

CC:

Test Plan:

Triggered file clustering and repartitioning, inspected that source_info.cpp was correctly updated.

Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

DiffCamp Revision: 123710