[Perf] Optimizing array lookup in APC
Summary:
Arrays in APC are represented not in their serialized form, but as key-value
pairs to avoid the high cost of unserialization, and are wrapped with SharedMap
in thread-local storage.

This diff divides the single ThreadSharedVariant map into two maps, one int64
map, and one StringData map. The StringData map is optimized to use the
precomputed keys, and to avoid all the unnecessary numeric comparisons. The
copy() method of StringData is also optimized to special-case on static
StringData.

It also fixes a bug in the escalation code of SharedMap, and adds a testcase
for it.

Test Plan:
fast_tests
slow_tests
production test

DiffCamp Revision: 127344
Reviewed By: myang
CC: hphp-diffs@lists, qixin, myang
Revert Plan:
OK