[Fix] In hphpi, Add extra incRef to files accessible under two paths
Summary:
If a file is stored under realpath(f) and f is included,
the table of files will have a new entry added under f to the file.
However, it was not incrementing the ref count of the file which meant
it would be decRef'd twice at the end of the request.

Test Plan:
sandbox

DiffCamp Revision: 124923
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK