fixed a re-declaring exception problem
Summary:
This is the bug that caused those platform unit tests to fail. This is related
to p_T not able to take DynamicObjectData. In this case, it has to use Object as
the type, so invokes are dynamically done.

Test Plan:
the new unit test, i'm testing www-dev now

DiffCamp Revision: 125189
Reviewed By: mwilliams
CC: hphp-diffs@lists, mwilliams
Revert Plan:
OK