Small fix to the stack translation problem
Summary:
The StackTrace(const std::string &) should be called for translation,
but const char* is directly passed in, which causes the StackTrace(bool)
is called instead.

This is the reason for always reporting the same StackTrace for leak
detector: The part to get the backtrace is correct but the part to
translate has problem.

Test Plan:
Test with hphpi leak detector and see it produce more meaningful
stack traces

DiffCamp Revision: 126082
Reviewed By: hzhao
Commenters: ervan
CC: hphp-diffs@lists, hzhao, qigao, ervan
Tasks:
211924: Xbox memory leak?

Revert Plan:
OK