Add a new list type
Summary:
Add a list type ListKindLeft
This behaves like ListKindComma, except that it evaluates to the left-most
element, rather than the right-most.
This is required for evaluating the return value of an inlined expression. The
value has to be evaluated, and then the local variables have to be unset. We
cant use a (php) variable to hold the return expression, because /that/ would
also need to be unset.
Its not yet used in the codebase. Just keeping the commits small.

Test Plan:
fast_tests, slow_tests, verify no code change for www.

DiffCamp Revision: 126664
Reviewed By: myang
CC: hphp-diffs@lists, myang
Revert Plan:
OK