[Fix] Dont try to parse new files after AnalyzeIncludes phase
Summary:
If an include's file-name expression was not initially scalar, but later gets
resolved, and the file had not already been parsed, we would attempt to parse it
during preOptimize, which doesnt work.
Change the behavior so that during preOptimize we can resolve an include to a
file that was parsed during the AnalyzeIncludes phase, but we never try to parse
a newly discovered file.

Test Plan:
build www-dev with --opts=all (this showed up as a crash building www-dev after
my recent include file changes, if --opts=all was also specified).

DiffCamp Revision: 128535
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK