[Fix] Better replicate php class declaration
Summary:
PHP is weird about class declaration. If a class has no parent or interfaces,
it is declared as soon as the file is included. If it has parents that are not
declared
in the file, but are defined when the file is included, it is defined as soon
as the file
is included. Otherwise, it is defined in its original position.

Test Plan:
TestCodeRunEval, sandbox, unit test fails in the same way.

DiffCamp Revision: 128705
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Tasks:
246372: Tests passing under hphpi but failing under www-dev

Revert Plan:
OK