[Fix] Ensure that builtin functions are visible when analyzing system/classes
Summary:
When system/classes is compiled, the builtin functions are all visible, so we
can use that information for type inference.
But when we are compiling something else, system/classes was being compiled
without that information. The result was that type inference could be wrong.

In the following, ReflectionClass::newInstance is implemented as returning an
object, because type inference worked when system/classes/reflectionclass.php
was compiled. But when this snippet was compiled, type inference said that it
returned a variant. The result was that the generated code didnt compile.

<?php
function test($className) {
  $x = new ReflectionClass($className);
  return $x->newInstance()->loadAll();
}

Test Plan:
fast_tests, slow_tests, the new test, flib/unit tests.

DiffCamp Revision: 129075
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK