Add unused flag to expression nodes
Summary:
There are a number of places where it would be convenient to know that an
expression is only being evaluated for its side effects. This adds a virtual
function to Construct which says for each child whether or not it uses the
result, and a pass in the AliasManager that updates a flag on Expression to say
whether or not its used.

Test Plan:
fast_tests, slow_tests

Nothing uses the flag yet, so not much to test.

DiffCamp Revision: 130435
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK