Eliminate locals by assignment propagation
Summary:
If an assignment can be moved to the point of first use of its variable, then
do so:

$a = foo();
echo $a;

becomes

echo ($a = foo())

If thats the only use of $a (and certain other conditions apply), we can then
eliminate the $a variable altogether.

This is mostly helpful for eliminating variables introduced by inlining.

Test Plan:
fast_tests, slow_tests, flib/unit tests (with --opts=all)

Observed a 6M reduction in stripped binary size for www-dev (when compiled with
--opts=all)

DiffCamp Revision: 131765
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao, mwilliams
Revert Plan:
OK