Fix incorrect type inference for ArrayPairExpression
Summary:
The keys were being coerced to Primitive type. This was normally harmless, but
broke this case:

<?php
class MyClass { function __toString() { return 'foo'; } }
$arr = array(new MyClass() => 2);
var_dump($arr);

In the above, when new MyClass() is coerced to Primitive, the resulting type is
String, which causes __toString to be called, which is wrong.
We have an almost identical test in test_code_run, except that it assigns the
new MyClass to a variable. Then because of some special casing in the type
inference for SimpleVariables, the problem doesnt show up.

Test Plan:
Extended the existing test case, fast_tests, slow_tests.

DiffCamp Revision: 129854
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK