eliminated an extra string copy from ostringstream to String
Summary:
This is an improvement I wanted to do at least 2 years ago :-)

1. Changed ExecutionContext::OutputBuffer to use StringBuffer, instead of
ostringstream, so we can have low level access.

2. Made all places using the StringBuffer not doing extra string copies. This
added a new StringBuffer::absorb() function for no-copy obFlush() if outer
buffer is empty.

3. Seperated ExecutionContext::obGetContents() into obCopyContents() and
obDetachContents(), so internally we don't have to make a copy if the buffer is
not needed afterwards for sure.

4. Removed StringBuffer::operator String(), which is ambiguous between copy()
and detach(). Now we have to be explicit about it.

Test Plan:
fast_tests, slow_tests, www

DiffCamp Revision: 131796
Reviewed By: mwilliams
CC: hphp-diffs@lists, mwilliams
Tasks:
259226: optimize output stream

Revert Plan:
OK