[Fix] possible StringData corruption in GlobalStatement
Summary:
The first parameter of flagGlobal() has type CStrRef, but in GlobalStatement,
the argument is n.data(), which causes a StringData object to be created with
AttachLiteral. This is not safe, as n itself might have been created with
CopyString. Then if n refers to the name of a dynamic variable, the key in the
underlying array of the LVariableTable would point to an invalid piece of
memory, after n is released.

Observed this corruption while debugging something else.

Test Plan:
made fast_tests

DiffCamp Revision: 137235
Reviewed By: iproctor
CC: hphp-diffs@lists, iproctor
Revert Plan:
OK