[Fix] Class constants bug when used in static initializer
Summary:
This is based off a patch submitted by lcastelli. His
description:
There is a bug when using an inherited class constants in a
class initializer. It is caused by a mismatch in the code that
determines whether to create a lazy static initializer, and the code the
decides whether the statement should go to the normal static initializer
or the lazy one. In particular the generation of the lazy static
initializer is forced if the value contains a dynamic static (see
class_variable.cpp
line 171), while the lazy static initializer is
used instead of the normal one in any case when the value is not scalar
(see class_variable.cpp lines 259 and 288).
  In the case of B::CONSTANT, it ends up being neither dynamic nor scalar,
so the lazy initializer is not generated and at the same time the
initialization is not performed in the normal initializer.
The difference between the behavior with A::CONSTANT and B::CONSTANT is
caused by the fact that in the pre-optimization phase the constant is
replaced by its value only if it's not inherited. This happens in
class_constant_expression.cpp, line 120, where
ConstantTable::isExplicitlyDeclared is called, which does not take into
consideration base classes. I'm not sure if this is really intended...

I made further changes so that nonexistent class constants were counted
as dynamic and so would be deferred to the lazy initializer and made the
check for declaration recursive.

Test Plan:
TestCodeRun
www

DiffCamp Revision: 138332
Reviewed By: mwilliams
CC: hphp-diffs@lists, macvicar, mwilliams
Revert Plan:
OK