[Fix] Deal with self when creating objects dynamically
Summary:
When the constructor invocation is invalid, a dynamic object
creation will be emitted. The name used is the original name, which
might be self or parent. This fix checks if it is self or parent and if
so uses the translated name instead.

Test Plan:
TestNewObjectExpression

DiffCamp Revision: 138968
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK