[Fix] Generate failed exit status when forked hphp dies unexpectedly
Summary:
When hphp forked a subprocess for the compilation step, and the subprocess was
terminated abnormally, the exit status check was incorrect.

Test Plan:
compiled some code with hphp, and killed the subprocess. Verified that hphp
reported failure, without trying to compile the generated files.

DiffCamp Revision: 140240
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK