Remove bad test case
Summary:
The new type.cpp caused this test to start failing. It doesn't actually appear
from
the language that this *should* fail, since the code in question is valid
PHP, unlike the other assertion in the same test case. Since the assertion
seems
bad, remove it.

(I should have checked this in with the rest of the type.cpp changes,
but forgot to do so.)

Test Plan:
fast_tests

DiffCamp Revision: 142035
Reviewed By: mwilliams
CC: hphp-diffs@lists, iproctor, mwilliams
Revert Plan:
OK