Horrible hack to mask slow_test loader failures
Summary:
Some of the slow tests fail some of the time with a loader error as they are
exiting. The tests have already completed at the time the error is emitted.
Pattern match the error, and ignore it.

It emits a warning in the output so we dont forget about the problem
altogether:

> TestObjectMethod passed
> runtime/tmp/Test0/main.php: Ignoring loader error: runtime/tmp/run:
>  symbol lookup error: /data/users/mwilliams/hphp-1/bin-O/libhphp_runtime.so:
>  undefined symbol: _ZN4HPHP14Synchronizable4waitEv
>
> TestClassMethod passed

Test Plan:
fast_tests, slow_tests

DiffCamp Revision: 142750
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao
Revert Plan:
OK