Ignore WarmupDocument under Sandbox
Summary:
It is not that easy to detect the interpreter mode, as we might be running in a
hybrid mode with both compiled and interpreted code. But we should be ignoring
WarmupDocument at least for sandbox, as explained in the comment.

Test Plan:
Started hphpi with SandboxMode=true, verified that WarmupDocument is ignored
for RPC.

DiffCamp Revision: 145127
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao, qixin, jxu, dhoskins
Tasks:
257646: HPHPi can't reset RPC thread

Revert Plan:
OK