Fix strrpos() and ob_end_clean() to match PHP
Summary:
The behaviour when a negative number is passed in is to start
at that position and work in that direction to 0.

For a positive number it starts at the end and works to the offset.

The PHP documentation is wrong here and doesn't match the behaviour that
PHP actualy does.

Also changed ob_end_clean() to return a value, this was causing problems
in the PHP test suite which used.

while(@ob_end_clean()); to remove all buffers.

Test Plan:
build
make fast_tests

run PHP tests which I've added

<?php
var_dump(strrpos("foof", "f", 1));

DiffCamp Revision: 145623
Reviewed By: hzhao
CC: hphp-diffs@lists, hzhao, macvicar
Revert Plan:
Ok